ceil gone

+++ libnd4j/tests_cpu/layers_tests/DeclarableOpsTests15.cpp
@@ -76,7 +76,7 @@ TEST_F(DeclarableOpsTests15, test_avgpooling_edge_1) {
     int inOutW = 35;
     int inOutC = 192;
 
-    auto x = NDArrayFactory::create<float>('c', {1, inOutH, inOutW, inOutC});
+    auto x = NDArrayFactory::create<double>('c', {1, inOutH, inOutW, inOutC});
     x.linspace(1.0);
 
     nd4j::ops::avgpool2d op;
@@ -91,8 +91,8 @@ TEST_F(DeclarableOpsTests15, test_avgpooling_edge_1) {
 
     int k = 3;
 
-    auto m = NDArrayFactory::create<float>('c', {1, inOutH, inOutW, inOutC});
-    auto c = NDArrayFactory::create<float>('c', {1, inOutH, inOutW, inOutC});
+    auto m = NDArrayFactory::create<double>('c', {1, inOutH, inOutW, inOutC});
+    auto c = NDArrayFactory::create<double>('c', {1, inOutH, inOutW, inOutC});
 
     for (int h = 0; h < inOutH; h++) {
         for (int w = 0; w < inOutW; w++) {
@@ -121,8 +121,8 @@ TEST_F(DeclarableOpsTests15, test_avgpooling_edge_1) {
                         idxIn[1] = kh;
                         idxIn[2] = kw;
 
-                        auto inVal = x.e<float>(0, kh, kw, ch);
-                        m.p(0, h, w, ch, inVal + m.e<float>(0, h, w, ch));
+                        auto inVal = x.e<double>(0, kh, kw, ch);
+                        m.p(0, h, w, ch, inVal + m.e<double>(0, h, w, ch));
                         c.p(0, h, w, ch, 1 + c.e<int>(0, h, w, ch));
                     }
                 }
